separated list storage
