+++ storage/memory/storage.go
@@ -161,28 +161,6 @@ func (s *storage) GetStringMap(key string) (map[string]string, error) {
 	return result, nil
 }
 
-func (s *storage) PopFromList(key string) (string, error) {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call PopFromList")
-
-	result, err := s.RedisStorage.PopFromList(key)
-	if err != nil {
-		return "", maskAny(err)
-	}
-
-	return result, nil
-}
-
-func (s *storage) PushToList(key string, element string) error {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call PushToList")
-
-	err := s.RedisStorage.PushToList(key, element)
-	if err != nil {
-		return maskAny(err)
-	}
-
-	return nil
-}
-
 func (s *storage) PushToSet(key string, element string) error {
 	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call PushToSet")
